% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankings.R
\name{fp_draft_rankings}
\alias{fp_draft_rankings}
\title{Fantasy Pros Expert Consensus Draft Rankings}
\usage{
fp_draft_rankings(pos = "overall", scoring = c("half", "ppr", "std"))
}
\arguments{
\item{pos}{\code{Charcater}. Specific position you want to return. Default will return overall.
Available options include
\itemize{
 \item \code{"overall"}
 \item \code{"QB"}
 \item \code{"RB"}
 \item \code{"WR"}
 \item \code{"TE"}
 \item \code{"K"}
 \item \code{"DST"}
 \item \code{"IDP"}
 \item \code{"DL"}
 \item \code{"LB"}
 \item \code{"DB"}
 \item \code{"Team QB"}
 \item \code{"Team RB"}
 \item \code{"Team WR"}
 \item \code{"Team TE"}
 \item \code{"Team K"}
 \item \code{"Team OL"}
 \item \code{"Head Coach"}
}}

\item{scoring}{\code{Charcater}. Fantasy scoring format. Default is \code{"half"}}
}
\value{
a tibble
}
\description{
Fantasy Pros Expert Consensus Draft Rankings
}
\examples{

# overall ECR for half point ppr
fp_draft_rankings()

# TE ECR using standard scoring
fp_draft_rankings(pos = "TE", scoring = "std")

# Individual Defensive Player ECR
fp_draft_rankings(pos = "IDP")
}
