% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snaps.R
\name{fp_snap_analysis}
\alias{fp_snap_analysis}
\title{Detailed Analysis of NFL snap counts}
\usage{
fp_snap_analysis(pos = "offense", season, start_week = 1,
  end_week = 17, scoring = c("half", "ppr", "std"))
}
\arguments{
\item{pos}{\code{Charcater}. Specific position you want to return. Default will return all
all offensive postions. Available options include
\itemize{
 \item \code{"offense"}
 \item \code{"QB"}
 \item \code{"RB"}
 \item \code{"WR"}
 \item \code{"TE"}
 \item \code{"Defense"}
 \item \code{"DL"}
 \item \code{"LB"}
 \item \code{"DB"}
}}

\item{season}{\code{Numeric}. The NFL season. If missing it will return the current year's season.
Supported season only go back to \code{2016}}

\item{start_week}{\code{Numeric}. The starting week. Default is \code{1}}

\item{end_week}{\code{Numeric}. The ending week. Default is \code{17}}
}
\value{
a tibble
}
\description{
Detailed Analysis of NFL snap counts
}
\note{
\itemize{
      \item If you query a single week all players that were on a bye that week are not returned
       }
}
\examples{

# all offensive positions for weeks 5-9 of the 2018 season
fp_snap_analysis(season = 2018, start_week = 5, end_week = 9)

# TE's in the 2017 season using standard scoring
fp_snap_analysis(pos = "TE", season = 2017, scoring = "std")

# all defensive positions for the current NFL season in week 1
fp_snap_analysis(pos = "defense", start_week = 1, end_week = start_week)
}
