% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snaps.R
\name{fp_snaps}
\alias{fp_snaps}
\title{NFL Snap Counts by Week}
\usage{
fp_snaps(pos = "offense", season, percentage = FALSE)
}
\arguments{
\item{pos}{\code{Charcater}. Specific position you want to return. Default will return all
all offensive postions. Available options include
\itemize{
 \item \code{"offense"}
 \item \code{"QB"}
 \item \code{"RB"}
 \item \code{"WR"}
 \item \code{"TE"}
 \item \code{"Defense"}
 \item \code{"DL"}
 \item \code{"LB"}
 \item \code{"DB"}
}}

\item{season}{\code{Numeric}. The NFL season. If missing it will return the current year's season.
Supported season only go back to \code{2016}}

\item{percentage}{\code{Logical}. Return counts as percentages? Default is \code{FALSE}}
}
\value{
a tibble
}
\description{
NFL Snap Counts by Week
}
\note{
\itemize{
       \item Due to how fantasypros returns the data, all player's teams are listed under their
      current team regardless of what season you query.
      \item Bye weeks are represented as \code{NA}
       }
}
\examples{

# all offensive positions for the 2018 season
fp_snaps(season = 2018)

# Percentage of total game snaps for TE's in the 2018 season
fp_snaps(pos = "TE", season = 2018, percentage = TRUE)

# all defensive positions for the current NFL season.
fp_snaps(pos = "defense")
}
