% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{fp_stats}
\alias{fp_stats}
\title{NFL Positional Statistics}
\usage{
fp_stats(pos = "QB", season, start_week = 1, end_week = 17,
  scoring = c("half", "ppr", "std"))
}
\arguments{
\item{pos}{\code{Charcater}. Specific position you want to return. Default will return all
all offensive postions. Available options include
\itemize{
 \item \code{"QB"}
 \item \code{"RB"}
 \item \code{"WR"}
 \item \code{"TE"}
 \item \code{"K"}
 \item \code{"DST"}
 \item \code{"DL"}
 \item \code{"LB"}
 \item \code{"DB"}
}}

\item{season}{\code{Numeric}. The NFL season. If missing it will return the current year's season.
Supported season only go back to \code{2015}}

\item{start_week}{\code{Numeric}. The starting week. Default is \code{1}}

\item{end_week}{\code{Numeric}. The ending week. Default is \code{17}}

\item{scoring}{\code{Charcater}. Fantasy scoring format. Default is \code{"half"}}
}
\value{
a tibble
}
\description{
NFL Positional Statistics
}
\note{
\itemize{
      \item If you query a single week all players that were on a bye that week are not returned
       }
}
\examples{

# overall fantasy points for the entire season
fp_stats("RB", season = 2018)

# TE's in the 2017 season using standard scoring
fp_stats(pos = "TE", season = 2017, scoring = "std")

}
