% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{fp_targets}
\alias{fp_targets}
\title{Target Counts by Week}
\usage{
fp_targets(pos = "offense", season)
}
\arguments{
\item{pos}{\code{Charcater}. Specific position you want to return. Default will return all
all offensive postions. Available options include
\itemize{
 \item \code{"offense"}
 \item \code{"RB"}
 \item \code{"WR"}
 \item \code{"TE"}
}}

\item{season}{\code{Numeric}. The NFL season. If missing it will return the current year's season.
Supported season only go back to \code{2013}}
}
\value{
a tibble
}
\description{
Target Counts by Week
}
\note{
\itemize{
      \item Bye weeks and games players did not play in are represented as \code{NA}
       }
}
\examples{

# all offensive positions for the 2018 season
fp_targets(season = 2018)

# total targets for TE's in the 2014 season
fp_targets(pos = "TE", season = 2014)

# all WR targets for the current NFL season.
fp_targets(pos = "WR")
}
