% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{fp_team_targets}
\alias{fp_team_targets}
\title{NFL team target distributions}
\usage{
fp_team_targets(season, start_week = 1, end_week = 17)
}
\arguments{
\item{season}{\code{Numeric}. The NFL season. If missing it will return the current year's season.
Supported season only go back to \code{2013}}

\item{start_week}{\code{Numeric}. The starting week. Default is \code{1}}

\item{end_week}{\code{Numeric}. The ending week. Default is \code{17}}
}
\value{
a tibble
}
\description{
NFL team target distributions
}
\examples{

fp_team_targets(season = 2018)

# total targets for TE's in the 2014 season
fp_team_targets(season = 2014, start_week = 8, end_week = 10)

}
